+++ lib/kibana/api.rb
@@ -1,0 +1,1 @@
+require_relative 'api/client'
