+++ lib/kibana/api_exceptions.rb
