+++ lib/kibana/http_status_codes.rb
