+++ lib/kibana/version.rb
@@ -1,0 +1,3 @@
+module Kibana
+  VERSION = "0.1.1"
+end
