--- lib/kibana_api/v7.rb
@@ -1,1 +1,0 @@
-require_relative 'v7/client'
