--- lib/kibana_api/v7/actions.rb
@@ -1,6 +1,0 @@
-require_relative 'actions/dashboard_actions'
-require_relative 'actions/index_pattern_actions'
-require_relative 'actions/saved_object_actions'
-require_relative 'actions/space_actions'
-require_relative 'actions/user_actions'
-require_relative 'actions/visualization_actions'
