--- lib/kibana_api/version.rb
@@ -1,3 +1,0 @@
-module KibanaAPI
-  VERSION = "0.1.0"
-end
