+++ spec/kibana_api_spec.rb
@@ -1,6 +1,6 @@
-RSpec.describe KibanaAPI do
+RSpec.describe Kibana do
   it "has a version number" do
-    expect(KibanaAPI::VERSION).not_to be nil
+    expect(Kibana::VERSION).not_to be nil
   end
 
   it "does something useful" do
